requirements.txt
psycopg2-binary
pandas
python-dotenv
matplotlib
seaborn